3.1.1 BUG FIX.
Introduced a bug in the latest commit.
At the line 253 of OBSERVABLES.f90 I introduced a code segment
which recalled the es_return_cvector(state_cvec) rather than
deallocate state_cvec itself.
This caused an error in the attempt of allocating an already
allocated array.

Fixed with the proper code segment
> if(allocated(state_cvec))deallocate(state_cvec)